geometrystream data can now be reserialized perfectly now. Changed the Padding attribute (int -> byte) and a little bit of string serializing to make use of this.

+++ libMBIN/Source/Template/NMSAttribute.cs
@@ -9,7 +9,7 @@ namespace libMBIN
         public bool Ignore { get; set; }
         public object DefaultValue { get; set; }
         public string[] EnumValue { get; set; }
-        public int Padding { get; set; }
+        public byte Padding { get; set; }
         public int Alignment { get; set; } = 4;
         public ulong GUID { get; set; }
         public bool Broken { get; set; }
